requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
django-request-logging==0.7.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==23.2
phonenumbers==8.13.30
phonenumberslite==8.13.30
pillow==10.2.0
proglog==0.1.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
